+++ src/Relude/Functor.hs
@@ -2,7 +2,7 @@
 
 {- |
 Copyright: (c) 2016 Stephen Diehl
-           (c) 20016-2018 Serokell
+           (c) 2016-2018 Serokell
            (c) 2018 Kowainik
 License:    MIT
 Maintainer: Kowainik <xrom.xkov@gmail.com>