+++ src/Relude/Lifted/Terminal.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE ExplicitForAll #-}
 
 {- |
-Copyright:  (c) 2018 Kowainik
+Copyright:  (c) 2018-2019 Kowainik
 License:    MIT
 Maintainer: Kowainik <xrom.xkov@gmail.com>