+++ src/Relude/Monad/Either.hs
@@ -4,9 +4,9 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 {- |
-Copyright: (c) 2016 Stephen Diehl
-           (c) 20016-2018 Serokell
-           (c) 2018 Kowainik
+Copyright:  (c) 2016 Stephen Diehl
+            (c) 2016-2018 Serokell
+            (c) 2018-2019 Kowainik
 License:    MIT
 Maintainer: Kowainik <xrom.xkov@gmail.com>