+++ CHANGELOG.md
@@ -4,9 +4,12 @@ Unreleased: 0.5.0
 =====
 
 * [#125](https://github.com/kowainik/relude/issues/125):
-  Moved many numerical functions and types in `Relude.Numeric`.    
-  Reexport `toIntegralSized` from `Data.Bits`.    
+  Moved many numerical functions and types in `Relude.Numeric`.
+  Reexport `toIntegralSized` from `Data.Bits`.
   Add `integerToBounded` and `integerToNatural` in `Relude.Numeric`.
+
+* [#129](https://github.com/kowainik/relude/issues/129):
+  Add `appliedTo` and `chainedTo` as named versions of operators `=<<` and `<**>`.
 
 0.4.0 — Nov 6, 2018
 =====